% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wald.R
\name{as.data.frame.wald}
\alias{as.data.frame.wald}
\title{Transform output of a Wald test into a data frame}
\usage{
\method{as.data.frame}{wald}(obj, se, digits = 3, sep = "", which = 1)
}
\arguments{
\item{obj}{}

\item{se}{}
}
\value{
a data frame with estimated coefficients, standard error and optionally, variables from a data frame.
}
\description{
Transform output of a Wald test into a data frame
}
\examples{
# coming soon
}
\author{
GM
}

