% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wald.R
\name{getX}
\alias{getX}
\title{Design matrix for a fit possibly on a new data frame}
\usage{
getX(fit, data = getData(fit))
}
\arguments{
\item{fit}{a fitted object with formula method}

\item{data}{(default NULL) a data frame on which to evaluate the design matrix}
}
\value{
a design matrix
}
\description{
This function return the X matrix for a fit possibly based on a different data frame than the model
}

