% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/long.R
\name{long}
\alias{long}
\title{Create a longitudinal file from a wide file}
\usage{
long(data, sep = "_", timevar = "time", idvar = "id",
  ids = 1:nrow(data), expand = TRUE, safe_sep = "#\%@!", ...)
}
\arguments{
\item{data}{wide data frame}

\item{sep}{(default '_') single character separator between long names and 'time' value.
Variable names with this separator are transformed to long variables.}

\item{timevar}{(default 'time') names of variable in long file to identify occasions. Its values are taken from the suffix following
the 'sep' character in each time-varying variable.}

\item{idvar}{(default: 'id') variable name used in output to identify row in input file. It may exist in the input (wide) file and must, in that case,
have a unique value in each row. If not, its values are replaced with row numbers.}

\item{ids}{(default \code{1:nrow(data)}) values for idvar in long file if the variable 'idvar' does not exist in in the wide file. Ignored
if \code{idvar} exists in \code{data}.}

\item{expand}{(default TRUE): if 'time' values are inconsistent, fill in missing 'time's with NAs.}

\item{safe_sep}{temporary safe? separator}
}
\value{
'long' file with each wide row repeated as many times as there are distinct values for the 'timevar' variable.
}
\description{
Uses a minimal set of arguments to create a long file using \link{\code{stats::reshape}}.
}
\examples{
z <- data.frame(id =letters[1:10], id2= 11:20, v_L = 1:10, v_R = 11:20)
long(z)
long(z, timevar = 'Side', idvar = 'idn', ids = LETTERS[1:10])
long(z, timevar = 'Side', idvar = 'idn', ids = z$id2)

# unbalanced times
z <- data.frame(id =letters[1:10], id2= 11:20, v_L = 1:10, v_R = 11:20, z_L = 21:30)
long(z)
}

