% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pch.R
\name{pch}
\alias{pch}
\title{Easy to remember pch characters}
\usage{
pch(shape = 1, fill = 1)
}
\arguments{
\item{shape}{square, circle, triangle, diamond, ltriangle (or 1,2,3,4,5)}

\item{fill}{empty, filled, filled with different color (or 1,2,3)}
}
\value{
value(s) of pch to produce corresponding character
}
\description{
Fifteen pch symbols can be generated with 5 different shapes and 3 different types of fill.
}

